+++ index.html
@@ -9,6 +9,8 @@
 
   <link rel="stylesheet" href="css/style.css" />
   <script src="scripts/main.js" type="module" defer></script>
+  <script src="scripts/menu.js" type="module" defer></script>
+  <script src="scripts/popup.js" type="module" defer></script>
   <script src="scripts/form.js" type="module" defer></script>
 </head>
